requirements.txt
pandas==1.2.3
finvizfinance==0.9.8
ta==0.7.0
Django==3.2
yfinance==0.1.55
djangorestframework==3.12.4
